pyproject.toml
[project]
name = "varscon-assessment"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Fadunsin"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "composio-langgraph (==0.7.8)",
    "langchain-openai (>=0.3.9,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
